+++ src/latents/__init__.py
@@ -4,22 +4,28 @@ import contextlib
 from importlib.metadata import PackageNotFoundError, version
 
 from latents import (
+    base,
+    callbacks,
     data,
     gfa,
     mdlag,
     observation,
     plotting,
     state,
+    tracking,
 )
 
 with contextlib.suppress(PackageNotFoundError):
     __version__ = version("latents")
 
 __all__ = [
+    "base",
+    "callbacks",
     "data",
     "gfa",
     "mdlag",
     "observation",
     "plotting",
     "state",
+    "tracking",
 ]